Adding KogRob & IPA Caller, IPA logo
