+++ pom.xml
@@ -33,7 +33,7 @@
 		<urcap.install.password>easybot</urcap.install.password>
 
 		<!--Install path for the UR Sim-->
-		<ursim.home>/home/ur/ursim/ursim-3.13.0.10253/</ursim.home>
+		<ursim.home>/home/chl/Software/ursim-5.8.2.10297</ursim.home>
 
 		<!--Host and standard user/password for UR Sim running in a VM-->
 		<ursimvm.install.host></ursimvm.install.host>
