+++ src/main/java/com/fzi/externalcontrol/impl/KogRobCaller.java
@@ -1,0 +1,12 @@
+package com.fzi.externalcontrol.impl;
+
+public class KogRobCaller extends IPACaller {
+
+  /*
+   * Default constructor
+   */
+	public KogRobCaller(String hostIp, String portNr, String xmlrpc_name) {
+		super(hostIp, portNr, xmlrpc_name);
+	}
+
+}
